move constants to stf root
