+++ src/stf/alg/circle.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // TODO (stouff) set this up with absolute includes
-#include "../math/constants.h"
+#include "../constants.h"
 #include "../math/vector.h"
 
 namespace stf {
@@ -10,10 +10,10 @@ namespace alg {
     template<typename T>
     inline T canonical_angle(T const theta)
     {
-        T canonical = std::fmod(theta, math::constants<T>::tau);
+        T canonical = std::fmod(theta, constants<T>::tau);
         if (canonical < T(0))
         {
-            canonical += math::constants<T>::tau;
+            canonical += constants<T>::tau;
         }
         return canonical;
     }
@@ -29,21 +29,21 @@ namespace alg {
         else
         {
             T tau = TWO_PI;
-            return (math::constants<T>::tau < phi) ? closest_equiv_angle(theta, phi - tau) : closest_equiv_angle(theta, phi + tau);
+            return (constants<T>::tau < phi) ? closest_equiv_angle(theta, phi - tau) : closest_equiv_angle(theta, phi + tau);
         }
     }
 
     template<typename T>
     inline T deg_to_rad(T const degrees)
     {
-        T constexpr convert = math::constants<T>::pi / T(180);
+        T constexpr convert = constants<T>::pi / T(180);
         return degrees * convert;
     }
 
     template<typename T>
     inline T rad_to_deg(T const radians)
     {
-        T constexpr convert = T(180) / math::constants<T>::pi;
+        T constexpr convert = T(180) / constants<T>::pi;
         return radians * convert;
     }
